ai_engine/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
google-generativeai==0.5.4
scikit-learn==1.4.2
beautifulsoup4==4.12.3
requests==2.31.0
python-multipart==0.0.22
pydantic==2.7.1
httpx==0.27.0
